+++ lib/src/maj_item_interface.dart
@@ -10,12 +10,17 @@
 import 'package:flutter/material.dart';
 import 'package:memory_and_json_directories/src/maj_node.dart';
 
+/// The interface used to ensure children of MAJNode have the correct functions
+/// it is recommended to add a static type name variable,but it is not required
+/// ex: static String typeName = "my_custom_item";
 abstract class MAJItemInterface {
   /// must return a string which will be a unique key in
   /// MAJNode.definitions map
   String getTypeName();
 
   /// return a widget which you wish to have displayed
+  /// nodeReference contains any data you would need
+  /// from a constructor
   Widget majBuild({
     required BuildContext context,
     required MAJNode nodeReference,